Added BaseNode class with ability to add Units to control who owns it
when a unit reaches the node it will trigger the addUnit function. This will first check if the state is neutral, it will then assign the state to the attacker. If the state is assigned, it will increase or decrease the number of units depending if the state and attacker match. Lastly if the units fall to zero it will reset the state to neutral